fix(index screen): prevent crash when pressing the favorites button twice quickly

+++ app/src/main/java/com/example/imnuricrestine/data/favorites/FavoritesLocalDataSource.java
@@ -37,7 +37,7 @@ public class FavoritesLocalDataSource implements FavoritesDataSource {
     @Override
     public CompletableFuture<Void> deleteFavorite(Favorite favorite) {
         return CompletableFuture.runAsync(() -> {
-            _favoritesDao.deleteFavorite(favorite);
+            if (favorite != null) _favoritesDao.deleteFavorite(favorite);
         });
     }
 }
